refactor: 예약 step3 activity fragment로 변경
