+++ app/src/main/res/layout/fragment_reservation_step3.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.reservation.ReservationStep3Activity">
+    tools:context=".ui.reservation.ReservationStep3Fragment">
 
     <ImageView
         android:id="@+id/back_btn"
